rename: Curricula를 Curriculum로 변경
