+++ src/main/java/com/twentyone/steachserver/domain/lecture/model/Lecture.java
@@ -1,6 +1,6 @@
 package com.twentyone.steachserver.domain.lecture.model;
 
-import com.twentyone.steachserver.domain.curricula.model.Curricula;
+import com.twentyone.steachserver.domain.curriculum.model.Curriculum;
 import com.twentyone.steachserver.domain.lectureStudents.model.LecturesStudents;
 import jakarta.persistence.*;
 import lombok.*;
@@ -29,7 +29,7 @@ public class Lecture {
 
     @ManyToOne
     @JoinColumn(name = "curriculum_id", nullable = false, referencedColumnName = "id")
-    private Curricula curricula;
+    private Curriculum curricula;
 
     @Column(name = "start_time", nullable = false)
     private LocalDateTime startTime;
